+++ CHANGELOG.md
@@ -3,6 +3,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) by Olivier Lacan, and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Added
+
+### Changed
+- Replaced `std::map` with `std::array` in `Fauna::ForageValues` to improve speed.
+
 ## 0.1.0 - 2019-11-07
 ### Added
 - Herbivores in cohort and individual mode.